requirements.txt
# =========================
# Core Networking & Scanning
# =========================
requests>=2.31.0
urllib3>=2.0.0
dnspython>=2.4.0
python-whois>=0.8.0

# =========================
# Port Scanning
# =========================
python-nmap>=0.7.1

# =========================
# SSL / Crypto Analysis
# =========================
pyOpenSSL>=23.2.0
cryptography>=41.0.0

# =========================
# Web Parsing / Scraping
# =========================
beautifulsoup4>=4.12.0
lxml>=4.9.0

# =========================
# Async / HTTP
# =========================
aiohttp>=3.8.0

# =========================
# API Server
# =========================
fastapi>=0.100.0
uvicorn>=0.23.0
pydantic>=2.0.0

# =========================
# ML / BERT URL Detection
# =========================
torch>=2.0.0
transformers>=4.38.0
sentencepiece>=0.1.99
huggingface-hub

# =========================
# Utilities
# =========================
tqdm>=4.65.0
colorama>=0.4.6
python-dotenv>=1.0.0